requirements.txt
groq
networkx
python-dotenv
matplotlib
